+++ src/compiler/scala/tools/nsc/javac/JavaScanners.scala
@@ -700,7 +700,7 @@ trait JavaScanners {
 
     def isSpecial(c: Char) = {
       val chtp = Character.getType(c)
-      chtp == Character.MATH_SYMBOL || chtp == Character.OTHER_SYMBOL
+      chtp == Character.MATH_SYMBOL.toInt || chtp == Character.OTHER_SYMBOL.toInt
     }
 
     private def getIdentRest {
