+++ src/compiler/scala/tools/nsc/symtab/StdNames.scala
@@ -113,7 +113,7 @@ trait StdNames {
            '|' | '\\'| '/' => true
       case _ =>
         val chtp = Character.getType(c)
-        chtp == Character.MATH_SYMBOL || chtp == Character.OTHER_SYMBOL
+        chtp == Character.MATH_SYMBOL.toInt || chtp == Character.OTHER_SYMBOL.toInt
       }
 
     /** If `name' is an expandedName name, the original name.
